pyproject.toml
[tool.poetry]
name = "social-network-drf"
version = "0.1.0"
description = "This is a test assignment from the StarNavi company"
authors = ["Chashin Efim <69835249+NONVME@users.noreply.github.com>"]
license = "MIT"
packages = [
  { include = "social_network" },
]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
python-dotenv = "^0.15.0"
djangorestframework = "^3.12.2"
coverage = "^5.5"

[tool.poetry.dev-dependencies]
bpython = "^0.21"
flake8 = "^3.9.0"
yapf = "^0.31.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
